Indicate mount points and mountable objects
This is done via an automatically-inserted {{mount}} template (there's a {{nomount}} companion to suppress the automatic insertion).

This is just a start, e.g. it might want to work more like {{reference}} and take an argument that can be incorporated into the generated text. Also, mount points could give a list of mountable objects, and mountable objects could give a list of mount points (these lists would be generated automatically by the report tool).