Improve mark_changed() logic so lspecs are only set to newer ones
The purpose of mark_changed() is to include parent nodes in "diffs" reports when a child (potentially far down the hierarchy) has changed. This was breaking for diagnostics where the command is new but the arguments are not (because they were originally defined as objects and parameters in CWMP diagnostics objects). The fix is more general (and might in fact fix other previously-unnoticed problems).