Add --altnotifreqstyle (alternative notification requirements style) option
This controls the generation of the 'Inform and Notification Requirements' section. The default is the existing CWMP style. If the option is specified, the section is called 'Notification Requirements', only the 'Parameters for which Active Notification MAY be Denied' sub-section is generated, and it's called 'Parameters for which Value Change Notification MAY be Denied'.