Support htmlbbf_deprecatedmodels option to mark deprecated models in the index file
The option value is a space-separated list of models (name and major version) that are deprecated. These are marked in the index file by (a) omitting them from the 'latest data models' list, (b) marking them '[DEPRECATED]' in the table of contents, (c) similarly marking them in the data model table.