Support Windows drive letters
Rather than ignoring the drive letter, the drive and directory are concatenated (so effectively drive+directory are treated as the directory). This hasn't been tested on a Windows system!

Note that very often, when extracting drive letter and directory from a path, they are ignored (only the filename is needed). Variable names have been changed to make this explicit.