Fix various problems with full XML (mostly related to commands and events)
Note: The feature/DMR-93-fix-problems-with-top-level-parameters-in-profiles branch was merged into this branch before making these changes (because the changes from both branches are needed in order to address all problems with full XML). Therefore the DMR-93 pull request should either be merged first or else declined. If it's merged first there shouldn't be any conflicts.

The changes are:
* When the {{enum}} template is used in an invalid context:
  * Avoid accessing the (undefined) list of enumeration values
  * Don't output {{enum}}'s "Enumeration of:" string
* For command / event argument parameters:
  * Convert object or factory defaults (defined in components) to parameter defaults
  * Ignore {{numentries}} templates (as above, will have come from a component)
  * In profiles, don't set the 'requirement' attribute

There are two places in the DEV2DM data model where {{enum}} is used in an invalid context. The above changes mean that these do no harm, but I'll create a DEV2DM issue anyway.