DMR-113: Change {{datatype}} macro to always include hyperlink to data type definition
The primary reason for doing this is so that it's clear (from the HTML) when a parameter uses a named data type.

There are a very small number of cases where parameter descriptions may need to be adjusted in order to read better. These can be addressed once the DMR PR has been merged.