DMR-139: Fix missing param (etc.) hyperlinks in object descriptions
The fix is a one-liner:
-    return ($name ? qq{''$name''} : $entity) unless $param;
+    return ($name ? qq{''$name''} : $entity) unless $object;

But fixing it exposed a long-standing problem in the Device:2 data model (possibly also others) where something like {{param||Enable}} has been written. That's a typo: the first argument (the parameter name) is empty and the second argument (the scope) is invalid. But it wasn't detected, because the empty parameter name immediately causes it to be treated the same as {{param}}. The fix exposes some confusing error messages where this happens, so I also added code to give a more helpful error message.

There are also some changes to the test support: I realised that the --diffs argument was missing when generating -full-diffs.html output. So I added it, and therefore also had to re-snapshot most of the expected test outputs.

And finally, there's a new test case for this issue.