DMR-115: Update 'find_file' logic to account for the desired amendment
This fixes a long-standing problem where searching for something like tr-181-2-common.xml would search for the highest matching amendment (c.f. the logic to search for the highest matching corrigendum). This is the wrong thing to do for amendments; the right thing is to check the actual amendments of candidate files (by reading their 'spec' attributes) and considering only matching files.

The fix also extends the spec_parse() routine to handle specs with omitted corrigenda, and uses the updated routine in several places where it wasn't previously possible to use it. I hope this hasn't broken anything...