DMR-194: Add {{deprecated}} etc. templates, with basic behavior
The templates are {{deprecated}}, {{obsoleted}} and {{deleted}} and they currently just output the requisite text without performing any additional checks.

References to deprecated, obsoleted or deleted parameters (or commands or events) are detected, but there's no check whether the reference is from a similarly deprecated, obsoleted or deleted item (which would be OK).