DMR-257: Add an integrate an XSD v1.1 validator
The validator (xmlschema-validate.py) uses the xmlschema python package
and is based on its xmlschema-validate utility.

The report tool collects all the XSD v1.1 files and then validates them
with a single invocation of the validator (this is much faster than
validating them one at a time, because the DM Schema only has to be
loaded once).

The levels of some of the report tool's debug messages have been adjusted.
Previously far too much was output at the lowest debug level. This makes
it feasible to use --loglevel debug to see the validator's debug output.

Note that the validator must be in the executable PATH. Validation can be
suppressed via a new --novalidate option.