DMR-278: Improve whitespace handling
The perl split() function was often used with the /\s+/ patterni
(one or more whitespace character) pattern, but this behaved badly
when there were leading space characters. All such cases have been
replaced with ' ' (single space), which is a special case that
ignores leading whitespace and otherwise behaves the same as /\s+/.

There are also a couple of minor unrelated changes that make the tool
a bit more resilient (these arose from testing the new tool).